+++ playground/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @byrding/playground
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @byrding/core@0.6.1
+  - @byrding/react@0.2.3
+  - @byrding/vue@0.2.3
+
 ## 0.0.6
 
 ### Patch Changes